Move AutoInitialize from common.py to MSVC/Util.py and fix log record prefix handling for debug output to stdout.
